% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebola}
\alias{ebola}
\title{ebola data}
\format{A dataframe
\describe{
   \item{id}{Unique patient identifier}
   \item{age}{Age in years}
   \item{age_group}{Grouped age}
   \item{sex}{Sex}
   \item{disease_onset}{Date of onset (date class)}
   \item{disease_ended}{Date of death or recovery (date class)}
   \item{days_at_risk}{Number of days at risk of death}
   \item{status}{Died or recovered (string)}
   \item{transmission}{How ebola was transmitted}
   \item{died}{Integer 1 = died, 0 = survived}
}}
\usage{
ebola
}
\description{
Person-time data for death from ebola. Data from
https://datacompass.lshtm.ac.uk/599/
 Marks, M. Learning Clinical Epidemiology with R. (Project). London School of Hygiene & Tropical Medicine, London, United Kingdom.
}
\keyword{datasets}
