% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tolbutamide}
\alias{tolbutamide}
\title{tolbutamide data - chapter 15-1, p260}
\format{An object of class \code{data.frame} with 409 rows and 3 columns.}
\usage{
tolbutamide
}
\description{
Age-specific comparison of death from all causes for Tolbutamide and placebo
treatment groups, University Group Diabetes Program (1970)
}
\details{
@format A dataframe with 409 observations and 3 variables
 \describe{
   \item{tolbutamide}{Given Tolbutamide (1) or placebo (0)}
   \item{dead}{Died (1) or surviving (0)}
   \item{age}{Less than 55 (<55) or 55 and over (ge55)}
 }
}
\keyword{datasets}
